Moved icons constants to specific file, added check when no resolver is assigned
