+++ ui/app/rif/constants/icons.js
@@ -1,0 +1,21 @@
+import { faCoins, faCheckCircle, faBolt, faArchive } from '@fortawesome/free-solid-svg-icons';
+
+export const DEFAULT_ICON = {
+  color: '#000080',
+  icon: faCoins,
+}
+
+export const domainIconProps = {
+  color: '#000080',
+  icon: faCheckCircle,
+}
+
+export const luminoNodeIconProps = {
+  color: '#508871',
+  icon: faBolt,
+}
+
+export const rifStorageIconProps = {
+  color: '#AD3232',
+  icon: faArchive,
+}
