+++ ui/app/rif/constants/index.js
@@ -1,25 +1,5 @@
-import { faCoins, faCheckCircle, faBolt, faArchive } from '@fortawesome/free-solid-svg-icons';
 import rifConfig from './../../../../rif.config';
-
-export const DEFAULT_ICON = {
-  color: '#000080',
-  icon: faCoins,
-}
-
-export const domainIconProps = {
-    color: '#000080',
-    icon: faCheckCircle,
-}
-
-export const luminoNodeIconProps = {
-    color: '#508871',
-    icon: faBolt,
-}
-
-export const rifStorageIconProps = {
-    color: '#AD3232',
-    icon: faArchive,
-}
+import { DEFAULT_ICON, domainIconProps, luminoNodeIconProps, rifStorageIconProps} from './icons';
 
 export const registrationTimeouts = {
   // number of seconds to wait before updating the page for the clock waiting.
@@ -33,4 +13,12 @@ export function GET_RESOLVERS () {
     name: 'Multi-Chain',
     address: rifConfig.rns.contracts.multiChainResolver,
   }]);
+}
+
+
+export {
+  DEFAULT_ICON,
+  domainIconProps,
+  luminoNodeIconProps,
+  rifStorageIconProps,
 }
