+++ ui/app/rif/pages/domainsDetailPage/domainDetailActive/domainDetailActive.js
@@ -4,7 +4,8 @@ import { connect } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
 import { getChainAddressByChainAddress } from '../../../utils/utils';
-import { CustomButton, AddNewChainAddressToResolver, DomainIcon, LuminoNodeIcon, RifStorageIcon, Menu } from '../../../components';
+import { CustomButton, DomainIcon, LuminoNodeIcon, RifStorageIcon, Menu } from '../../../components';
+import AddNewChainAddressToResolver from './addNewTokenNetworkAddress/addNewChainAddressToResolver';
 import { GET_RESOLVERS, DEFAULT_ICON } from '../../../constants';
 import { SLIP_ADDRESSES } from '../../../constants/slipAddresses';
 import niftyActions from '../../../../actions';
